[CSL-2328] getBlocks takes tip hash not header
All the info in block header was redundant, we cared only about the
hash.

NB this patch does not eliminate the waste of calling 'headerHash' on
the block header, which was done anyway. In block retrieval, this was
done for every one of the headers received from the server: they were
reserialized and hashed.